nettoyage_express_ancien/deliverable/src/netexpress/requirements/dev.txt
-r base.txt
pytest>=8.3
pytest-django>=4.8